Add and rearrange tests.
